Update
major update now student can view AI generated motivational message
