+++ frontEnd/package.json
@@ -16,12 +16,14 @@
     "@fullcalendar/react": "^6.1.15",
     "@fullcalendar/timegrid": "^6.1.15",
     "axios": "^1.7.7",
+    "bootstrap": "^5.3.3",
     "cors": "^2.8.5",
     "lucide-react": "^0.453.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icon": "^1.0.0",
     "react-icons": "^5.3.0",
+    "react-markdown": "^9.0.1",
     "react-router-dom": "^6.26.2",
     "xlsx": "^0.18.5"
   },
